requirements.txt
beautifulsoup4==4.12.2
celery==5.3.1
Django==4.2.3
djangorestframework==3.14.0
google-play-scraper==1.2.4
requests==2.31.0
redis==4.6.0